Removed unused react logo, created componets folder and moved components there, moved all css to styles folder, updated path to moved components/styles
