+++ src/components/fusTheme.js
@@ -1,0 +1,39 @@
+/**
+ * More info here: http://www.material-ui.com/#/customization/themes
+ */
+
+import {
+  cyan500,
+  cyan700,
+  pinkA200,
+  grey100,
+  grey300,
+  grey400,
+  grey500,
+  white,
+  darkBlack,
+  fullBlack
+} from 'material-ui/styles/colors'
+import { fade } from 'material-ui/utils/colorManipulator'
+import spacing from 'material-ui/styles/spacing'
+
+export default {
+  spacing: spacing,
+  fontFamily: 'Roboto, sans-serif',
+  palette: {
+    primary1Color: '#21412a',
+    primary2Color: cyan700,
+    primary3Color: grey400,
+    accent1Color: pinkA200,
+    accent2Color: grey100,
+    accent3Color: grey500,
+    textColor: darkBlack,
+    alternateTextColor: white,
+    canvasColor: white,
+    borderColor: grey300,
+    disabledColor: fade(darkBlack, 0.3),
+    pickerHeaderColor: cyan500,
+    clockCircleColor: fade(darkBlack, 0.07),
+    shadowColor: fullBlack
+  }
+}
