+++ src/views/Letterhead.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import LetterheadCard from '../LetterheadCard'
+import LetterheadCard from '../components/LetterheadCard'
 
 class Letterhead extends Component {
   render () {
